move operator metric gauge
